+++ .eslintrc.js
@@ -20,6 +20,7 @@ module.exports = {
     'prettier/prettier': 0,
     'no-unused-vars': 'off',
     '@typescript-eslint/no-unused-vars': ['error'],
+    'react/prop-types': 'off',
   },
   settings: {
     react: {
