+++ features/scheduler/create-schedule-form/step-2.tsx
@@ -1,5 +1,15 @@
-import { format, getHours, parse } from 'date-fns'
-import { useState } from 'react'
+import { addWeeks, format, getHours, parse, setDay } from 'date-fns'
+import {
+  CreateScheduleMutationVariables,
+  Schedule_Day_Week_Day_Enum,
+  useCreateScheduleMutation,
+} from 'generated/graphql'
+import { ClientError } from 'graphql-request'
+import useOnLeave from 'hooks/use-on-leave'
+import { useRouter } from 'next/router'
+import { FormEvent, useState } from 'react'
+import { useToasts } from 'react-toast-notifications'
+import monthsMappings from 'utils/months-mappings'
 import { FormData } from './step-1'
 import StepTwoNewTask, { NewTask } from './step-2-new-task'
 
@@ -54,10 +64,25 @@ type TimeIconProps = {
   className?: string
 }
 
-function NewScheduleFormStepTwo({
+function CreateScheduleFormStepTwo({
   previousStepData,
-}: NewScheduleFormStepTwoProps) {
-  const [isNewTaskOpen, setIsNewTaskOpen] = useState(false)
+}: CreateScheduleFormStepTwoProps) {
+  const { addToast } = useToasts()
+  const router = useRouter()
+  const { mutate, status } = useCreateScheduleMutation<ClientError>({
+    onError: error => {
+      addToast(
+        <h3 className="text-sm font-medium text-red-800">{error.message}</h3>,
+        { appearance: 'error' },
+      )
+    },
+    onSuccess: data => {
+      const newScheduleId = data.insert_schedule?.returning[0].id
+      router.push(`/scheduler/${newScheduleId}`)
+    },
+  })
+
+  const { ref, isVisible, setIsVisible } = useOnLeave(false)
   const weekDays = Object.keys(previousStepData.days)
     .filter(key => previousStepData.days[key])
     .map(capitalizeFirstLetter)
@@ -108,27 +133,59 @@ function NewScheduleFormStepTwo({
     })
   }
 
+  const submit = (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault()
+
+    const variables: CreateScheduleMutationVariables = {
+      title: previousStepData.title,
+      days: Object.keys(weekTasks).map(day => {
+        const validWeekDay = day.toLowerCase()
+        return {
+          week_day: validWeekDay as Schedule_Day_Week_Day_Enum,
+          tasks: {
+            data: weekTasks[day].map(task => {
+              const parsed = parse(task.startTime, 'HH:mm', new Date())
+
+              const startTime = addWeeks(
+                setDay(parsed, monthsMappings[day]),
+                1,
+              ).toISOString()
+
+              return {
+                description: task.description,
+                start_time: startTime,
+              }
+            }),
+          },
+        }
+      }),
+    }
+
+    mutate(variables)
+  }
+
   return (
     <>
       <h2 className="max-w-6xl items-center mb-2 mx-auto mt-8 text-lg leading-6 font-medium text-gray-900">
         Add tasks to your new schedule or skip this and do it later.
       </h2>
-      <p className="text-base mb-2 leading-6 font-medium text-gray-700">
+      <p className="text-base sm:mb-2 mb-3 leading-6 font-medium text-gray-700">
         Tasks are events or actions such as walking the dog, taking out the
         trash and anything else you can think of. They make up the days of a
         schedule and are essential for making your schedules effective.
       </p>
       <div className="relative">
         <button
-          onClick={() => setIsNewTaskOpen(isOpenOrClosed => !isOpenOrClosed)}
+          onClick={() => setIsVisible(true)}
           className="inline-flex items-center px-2.5 py-1.5 border border-transparent text-sm font-medium rounded text-indigo-700 bg-indigo-100 hover:bg-indigo-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
         >
           New task
         </button>
         <StepTwoNewTask
+          ref={ref}
           onSubmit={addTask}
-          onClose={() => setIsNewTaskOpen(false)}
-          isOpen={isNewTaskOpen}
+          onClose={() => setIsVisible(false)}
+          isOpen={isVisible}
           weekDays={weekDays}
         />
       </div>
@@ -144,9 +201,9 @@ function NewScheduleFormStepTwo({
                   {weekDay}
                 </span>
               </div>
-              <div className="mt-4 sm:mt-0 sm:col-span-2">
+              <div className="sm:col-span-2">
                 <div className="flow-root">
-                  <ul className="-mb-8">
+                  <ul className="sm:-mb-8">
                     {weekTasks[weekDay].length > 0 ? (
                       weekTasks[weekDay].map((task, index, weekTasksRef) => (
                         <li key={task.id}>
@@ -194,7 +251,7 @@ function NewScheduleFormStepTwo({
                       ))
                     ) : (
                       <li>
-                        <div className="min-w-0 flex-1 pt-2.5 flex justify-between space-x-4">
+                        <div className="min-w-0 flex-1 sm:pt-2.5 flex justify-between space-x-4">
                           <p className="text-sm text-gray-500">
                             No tasks for this day
                           </p>
@@ -207,18 +264,53 @@ function NewScheduleFormStepTwo({
             </div>
           ))}
         </div>
-        <form>
-          <div className="flex mt-12">
+        <form onSubmit={submit}>
+          <div className="flex mt-4 sm:mt-12">
             <button
+              onClick={() => {
+                mutate({
+                  title: previousStepData.title,
+                  days: Object.keys(weekTasks).map(day => {
+                    const validWeekDay = day.toLowerCase()
+                    return {
+                      week_day: validWeekDay as Schedule_Day_Week_Day_Enum,
+                    }
+                  }),
+                })
+              }}
               type="button"
-              className="inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+              disabled={status === 'loading'}
+              className="inline-flex items-center disabled:cursor-not-allowed disabled:opacity-60 px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
             >
               Skip
             </button>
             <button
-              type="button"
-              className="ml-2 inline-flex items-center px-4 py-2 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+              type="submit"
+              disabled={status === 'loading'}
+              className="ml-2 inline-flex disabled:cursor-not-allowed disabled:opacity-60 items-center px-4 py-2 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
             >
+              {status === 'loading' && (
+                <svg
+                  className="animate-spin -ml-1 mr-3 h-5 w-5 text-white"
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                >
+                  <circle
+                    className="opacity-25"
+                    cx="12"
+                    cy="12"
+                    r="10"
+                    stroke="currentColor"
+                    strokeWidth="4"
+                  ></circle>
+                  <path
+                    className="opacity-75"
+                    fill="currentColor"
+                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+                  ></path>
+                </svg>
+              )}
               Continue
             </button>
           </div>
@@ -236,8 +328,8 @@ type Task = {
   startTime: string
 }
 
-type NewScheduleFormStepTwoProps = {
+type CreateScheduleFormStepTwoProps = {
   previousStepData: FormData
 }
 
-export default NewScheduleFormStepTwo
+export default CreateScheduleFormStepTwo
