+++ hasura/metadata/query_collections.yaml
@@ -27,5 +27,5 @@
         query:
           "mutation CreateSchedule($title: String!, $days:
           [schedule_day_insert_input!]!) {\r\n  insert_schedule(objects: {
-          title: $title, days: { data: $days } })
-          {\r\n    affected_rows\r\n  }\r\n}"
+          title: $title, days: { data: $days } }) {\r\n    returning
+          {\r\n      id\r\n    }\r\n  }\r\n}"
