+++ utils/months-mappings.ts
@@ -1,0 +1,9 @@
+export default {
+  Monday: 1,
+  Tuesday: 2,
+  Wednesday: 3,
+  Thursday: 4,
+  Friday: 5,
+  Saturday: 6,
+  Sunday: 0,
+}
